Move memory limiter from the `Resource\Limiter` namespace to `Resource\Memory`
