+++ src/Container.php
@@ -74,8 +74,8 @@ use Infection\Process\Runner\InitialTestsRunner;
 use Infection\Process\Runner\MutationTestingRunner;
 use Infection\Process\Runner\Parallel\ParallelProcessRunner;
 use Infection\Process\Runner\TestRunConstraintChecker;
-use Infection\Resource\Limiter\MemoryLimiter;
 use Infection\Resource\Memory\MemoryFormatter;
+use Infection\Resource\Memory\MemoryLimiter;
 use Infection\Resource\Time\Stopwatch;
 use Infection\Resource\Time\TimeFormatter;
 use Infection\TestFramework\CommandLineBuilder;
